currency system with buyable characters
currency is working and being displayed on screen. money is currently being generated when walking, for testing. also you can buy different characters, but it has the same problems as character switching script: text doesn't appear when walking from below, sometimes it appears dim. when character is bought you have to leave and enter coalisions to interact with it again.